+++ CHANGELOG.md
@@ -11,4 +11,4 @@
 
 ## [0.1.0]
 
-- Initial release
+-   Initial release